Replaces BigInteger with FieldElement
Introduces a framework to explicitly define and represent the field in which computation is being done in arithmetic protocols. This allows us to encapsulate the internal representation of the field elements (which were previously just represented as raw BigInteger's) and how computations are done on them, which makes it easier for us to experiment with different approaches.

To illustrate this benefit we include two implementations, one that is simply a port of the old method naively using the BigInteger class, and one based on fields with a pseudo Mersenne prime modulus which allows for faster modular arithmetic.

Unrelated to the above, this PR also removes the AsyncNetwork class, and replaces it with the SocketNetwork class. This is done as the AsyncNetwork implementation was not working on Windows.